requirements.txt
torch
torchvision 
opencv-python 
matplotlib 
pillow 
numpy
timm